Extract common SMART SL code into a module
